@@ -10,57 +10,64 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "postinstall": "husky install && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
+    "postinstall": "husky install"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.7",
-    "@angular/cdk": "^12.2.7",
-    "@angular/common": "^12.2.7",
-    "@angular/compiler": "^12.2.7",
-    "@angular/core": "^12.2.7",
-    "@angular/forms": "^12.2.7",
-    "@angular/material": "^12.2.7",
-    "@angular/platform-browser": "^12.2.7",
-    "@angular/platform-browser-dynamic": "^12.2.7",
-    "@angular/router": "^12.2.7",
-    "bcrypt": "^3.0.2",
+    "@angular/animations": "^13.0.2",
+    "@angular/cdk": "^13.0.2",
+    "@angular/common": "^13.0.2",
+    "@angular/compiler": "^13.0.2",
+    "@angular/core": "^13.0.2",
+    "@angular/forms": "^13.0.2",
+    "@angular/material": "^13.0.2",
+    "@angular/platform-browser": "^13.0.2",
+    "@angular/platform-browser-dynamic": "^13.0.2",
+    "@angular/router": "^13.0.2",
+    "bcrypt": "^5.0.1",
     "body-parser": "^1.18.2",
     "compression": "^1.7.2",
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.4",
-    "dotenv": "^6.0.0",
+    "dotenv": "^10.0.0",
     "events": "^3.0.0",
     "express": "^4.16.3",
     "express-async-handler": "^1.1.3",
     "express-jwt": "^5.3.1",
     "express-validation": "^1.0.2",
-    "formidable": "^1.2.1",
-    "helmet": "^3.21.1",
+    "formidable": "^3.1.3",
+    "helmet": "^4.6.0",
     "http-errors": "^1.6.3",
-    "joi": "^13.3.0",
+    "joi": "^17.4.2",
     "jsonwebtoken": "^8.2.1",
-    "method-override": "^2.3.10",
-    "mongoose": "^5.7.5",
+    "method-override": "^3.0.0",
+    "mongoose": "^6.0.13",
     "morgan": "^1.9.1",
-    "nodemon": "^1.17.5",
+    "nodemon": "^2.0.15",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "rxjs": "^7.0.1",
-    "swagger-ui-express": "^3.0.9",
+    "rxjs": "^7.4.0",
+    "swagger-ui-express": "^4.1.6",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.7",
-    "@angular/cli": "^12.2.7",
-    "@angular/compiler-cli": "^12.2.7",
-    "@angular/language-service": "^12.2.7",
+    "@angular-devkit/build-angular": "~13.0.3",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/cli": "^13.0.3",
+    "@angular/compiler-cli": "^13.0.2",
+    "@angular/language-service": "^13.0.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~15.0.0",
-    "codelyzer": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "5.3.0",
+    "@typescript-eslint/parser": "5.3.0",
     "concurrently": "^3.5.1",
+    "eslint": "^8.2.0",
     "husky": "^7.0.2",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -72,12 +79,6 @@
     "lint-staged": "^11.1.2",
     "prettier": "^2.4.1",
     "ts-node": "~6.1.0",
-    "tslint": "~6.1.0",
-    "typescript": "4.2.4"
-  },
-  "lint-staged": {
-    "*.{js,ts,html,scss}": [
-      "prettier --write"
-    ]
+    "typescript": "4.4.4"
   }
 }
