@@ -18,7 +18,12 @@ export function appInitializerFactory(authService: AuthService) {
 }
 
 @NgModule({
-  imports: [BrowserAnimationsModule, HttpClientModule, SharedModule, AppRoutingModule],
+  imports: [
+    BrowserAnimationsModule,
+    HttpClientModule,
+    SharedModule,
+    AppRoutingModule,
+  ],
   declarations: [AppComponent, HeaderComponent, HomeComponent],
   providers: [
     {
