@@ -6,26 +6,24 @@ require('dotenv').config();
 // define validation for all the env vars
 const envVarsSchema = Joi.object({
   NODE_ENV: Joi.string()
-    .allow(['development', 'production', 'test', 'provision'])
+    .allow('development', 'production', 'test', 'provision')
     .default('development'),
-  SERVER_PORT: Joi.number()
-    .default(4040),
-  MONGOOSE_DEBUG: Joi.boolean()
-    .when('NODE_ENV', {
-      is: Joi.string().equal('development'),
-      then: Joi.boolean().default(true),
-      otherwise: Joi.boolean().default(false)
-    }),
-  JWT_SECRET: Joi.string().required()
+  SERVER_PORT: Joi.number().default(4040),
+  MONGOOSE_DEBUG: Joi.boolean().when('NODE_ENV', {
+    is: Joi.string().equal('development'),
+    then: Joi.boolean().default(true),
+    otherwise: Joi.boolean().default(false),
+  }),
+  JWT_SECRET: Joi.string()
+    .required()
     .description('JWT Secret required to sign'),
-  MONGO_HOST: Joi.string().required()
-    .description('Mongo DB host url'),
-  MONGO_PORT: Joi.number()
-    .default(27017)
-}).unknown()
+  MONGO_HOST: Joi.string().required().description('Mongo DB host url'),
+  MONGO_PORT: Joi.number().default(27017),
+})
+  .unknown()
   .required();
 
-const { error, value: envVars } = Joi.validate(process.env, envVarsSchema);
+const { error, value: envVars } = envVarsSchema.validate(process.env);
 if (error) {
   throw new Error(`Config validation error: ${error.message}`);
 }
@@ -38,8 +36,8 @@ const config = {
   frontend: envVars.MEAN_FRONTEND || 'angular',
   mongo: {
     host: envVars.MONGO_HOST,
-    port: envVars.MONGO_PORT
-  }
+    port: envVars.MONGO_PORT,
+  },
 };
 
 module.exports = config;
