@@ -12,7 +12,7 @@ const swaggerUi = require('swagger-ui-express');
 const swaggerDocument = require('./swagger.json');
 const routes = require('../routes/index.route');
 const config = require('./config');
-const passport = require('./passport')
+const passport = require('./passport');
 
 const app = express();
 
@@ -22,25 +22,28 @@ if (config.env === 'development') {
 
 // Choose what fronten framework to serve the dist from
 var distDir = '../../dist/';
-if (config.frontend == 'react'){
-  distDir ='../../node_modules/material-dashboard-react/dist'
- }else{
-  distDir ='../../dist/' ;
- }
+if (config.frontend == 'react') {
+  distDir = '../../node_modules/material-dashboard-react/dist';
+} else {
+  distDir = '../../dist/';
+}
 
-// 
-app.use(express.static(path.join(__dirname, distDir)))
+//
+app.use(express.static(path.join(__dirname, distDir)));
 app.use(/^((?!(api)).)*/, (req, res) => {
   res.sendFile(path.join(__dirname, distDir + '/index.html'));
 });
 
 console.log(distDir);
- //React server
-app.use(express.static(path.join(__dirname, '../../node_modules/material-dashboard-react/dist')))
+//React server
+app.use(
+  express.static(
+    path.join(__dirname, '../../node_modules/material-dashboard-react/dist')
+  )
+);
 app.use(/^((?!(api)).)*/, (req, res) => {
-res.sendFile(path.join(__dirname, '../../dist/index.html'));
-}); 
-
+  res.sendFile(path.join(__dirname, '../../dist/index.html'));
+});
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -64,21 +67,20 @@ app.use('/api/', routes);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
-  const err = new httpError(404)
+  const err = new httpError(404);
   return next(err);
 });
 
 // error handler, send stacktrace only during development
 app.use((err, req, res, next) => {
-
   // customize Joi validation errors
   if (err.isJoi) {
-    err.message = err.details.map(e => e.message).join("; ");
+    err.message = err.details.map(e => e.message).join('; ');
     err.status = 400;
   }
 
   res.status(err.status || 500).json({
-    message: err.message
+    message: err.message,
   });
   next(err);
 });
