@@ -6,12 +6,7 @@ const config = require('./config');
 
 // connect to mongo db
 const mongoUri = config.mongo.host;
-mongoose.connect(mongoUri, { 
-  keepAlive: 1,
-  useCreateIndex: true,
-  useNewUrlParser: true, 
-  useUnifiedTopology: true
-});
+mongoose.connect(mongoUri, { keepAlive: 1 });
 mongoose.connection.on('error', () => {
   throw new Error(`unable to connect to database: ${mongoUri}`);
 });
@@ -22,4 +17,3 @@ if (config.MONGOOSE_DEBUG) {
     debug(`${collectionName}.${method}`, util.inspect(query, false, 20), doc);
   });
 }
-
