@@ -7,30 +7,38 @@ const bcrypt = require('bcrypt');
 const User = require('../models/user.model');
 const config = require('./config');
 
-const localLogin = new LocalStrategy({
-  usernameField: 'email'
-}, async (email, password, done) => {
-  let user = await User.findOne({ email });
-  if (!user || !bcrypt.compareSync(password, user.hashedPassword)) {
-    return done(null, false, { error: 'Your login details could not be verified. Please try again.' });
+const localLogin = new LocalStrategy(
+  {
+    usernameField: 'email',
+  },
+  async (email, password, done) => {
+    let user = await User.findOne({ email });
+    if (!user || !bcrypt.compareSync(password, user.hashedPassword)) {
+      return done(null, false, {
+        error: 'Your login details could not be verified. Please try again.',
+      });
+    }
+    user = user.toObject();
+    delete user.hashedPassword;
+    done(null, user);
   }
-  user = user.toObject();
-  delete user.hashedPassword;
-  done(null, user);
-});
+);
 
-const jwtLogin = new JwtStrategy({
-  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-  secretOrKey: config.jwtSecret
-}, async (payload, done) => {
-  let user = await User.findById(payload._id);
-  if (!user) {
-    return done(null, false);
+const jwtLogin = new JwtStrategy(
+  {
+    jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+    secretOrKey: config.jwtSecret,
+  },
+  async (payload, done) => {
+    let user = await User.findById(payload._id);
+    if (!user) {
+      return done(null, false);
+    }
+    user = user.toObject();
+    delete user.hashedPassword;
+    done(null, user);
   }
-  user = user.toObject();
-  delete user.hashedPassword;
-  done(null, user);
-});
+);
 
 passport.use(jwtLogin);
 passport.use(localLogin);
