@@ -7,16 +7,15 @@ const userSchema = Joi.object({
   email: Joi.string().email(),
   mobileNumber: Joi.string().regex(/^[1-9][0-9]{9}$/),
   password: Joi.string().required(),
-  repeatPassword: Joi.string().required().valid(Joi.ref('password'))
-})
-
+  repeatPassword: Joi.string().required().valid(Joi.ref('password')),
+});
 
 module.exports = {
-  insert
-}
+  insert,
+};
 
 async function insert(user) {
-  user = await Joi.validate(user, userSchema, { abortEarly: false });
+  user = await userSchema.validateAsync(user, { abortEarly: false });
   user.hashedPassword = bcrypt.hashSync(user.password, 10);
   delete user.password;
   return await new User(user).save();
