@@ -1,5 +1,5 @@
 const express = require('express');
-const asyncHandler = require('express-async-handler')
+const asyncHandler = require('express-async-handler');
 const passport = require('passport');
 const userCtrl = require('../controllers/user.controller');
 const authCtrl = require('../controllers/auth.controller');
@@ -9,16 +9,19 @@ const router = express.Router();
 module.exports = router;
 
 router.post('/register', asyncHandler(register), login);
-router.post('/login', passport.authenticate('local', { session: false }), login);
+router.post(
+  '/login',
+  passport.authenticate('local', { session: false }),
+  login
+);
 router.get('/me', passport.authenticate('jwt', { session: false }), login);
-
 
 async function register(req, res, next) {
   let user = await userCtrl.insert(req.body);
   user = user.toObject();
   delete user.hashedPassword;
   req.user = user;
-  next()
+  next();
 }
 
 function login(req, res) {
