@@ -5,9 +5,7 @@ const authRoutes = require('./auth.route');
 const router = express.Router(); // eslint-disable-line new-cap
 
 /** GET /health-check - Check service health */
-router.get('/health-check', (req, res) =>
-  res.send('OK')
-);
+router.get('/health-check', (req, res) => res.send('OK'));
 
 router.use('/auth', authRoutes);
 router.use('/user', userRoutes);
