@@ -6,11 +6,9 @@ const userCtrl = require('../controllers/user.controller');
 const router = express.Router();
 module.exports = router;
 
-router.use(passport.authenticate('jwt', { session: false }))
+router.use(passport.authenticate('jwt', { session: false }));
 
-router.route('/')
-  .post(asyncHandler(insert));
-
+router.route('/').post(asyncHandler(insert));
 
 async function insert(req, res) {
   let user = await userCtrl.insert(req.body);
