@@ -4,18 +4,21 @@ import { RouterModule, Routes } from '@angular/router';
 import { AdminComponent } from './admin.component';
 import { OnlyAdminUsersGuard } from './admin-user-guard';
 
-const routes: Routes = [{
-  path: 'admin',
-  canActivate: [OnlyAdminUsersGuard],
-  children: [{
-    path: '',
-    component: AdminComponent,
-  }]
-}];
+const routes: Routes = [
+  {
+    path: 'admin',
+    canActivate: [OnlyAdminUsersGuard],
+    children: [
+      {
+        path: '',
+        component: AdminComponent,
+      },
+    ],
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-
 export class AdminRoutingModule {}
