@@ -1,15 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-admin',
   templateUrl: './admin.component.html',
 })
-export class AdminComponent implements OnInit {
-
-  constructor() {}
-
-  public ngOnInit() {
-  }
-}
-
-
+export class AdminComponent {}
