@@ -1,19 +1,13 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
 import { AdminRoutingModule } from './admin-routing.module';
-import {AdminComponent} from './admin.component';
-import {OnlyAdminUsersGuard} from './admin-user-guard';
+import { AdminComponent } from './admin.component';
+import { OnlyAdminUsersGuard } from './admin-user-guard';
 
 @NgModule({
-  declarations: [
-    AdminComponent
-  ],
-  imports: [
-    CommonModule,
-    AdminRoutingModule,
-  ],
-  providers: [
-    OnlyAdminUsersGuard
-  ]})
+  declarations: [AdminComponent],
+  imports: [CommonModule, AdminRoutingModule],
+  providers: [OnlyAdminUsersGuard],
+})
 export class AdminModule {}
