@@ -31,7 +31,10 @@ export class RegisterComponent {
     fullname: new FormControl('', [Validators.required]),
     email: new FormControl('', [Validators.required, Validators.email]),
     password: new FormControl('', [Validators.required]),
-    repeatPassword: new FormControl('', [Validators.required, this.passwordsMatchValidator]),
+    repeatPassword: new FormControl('', [
+      Validators.required,
+      this.passwordsMatchValidator,
+    ]),
   });
 
   get fullname(): AbstractControl {
@@ -55,10 +58,13 @@ export class RegisterComponent {
       return;
     }
 
-    const { fullname, email, password, repeatPassword } = this.userForm.getRawValue();
+    const { fullname, email, password, repeatPassword } =
+      this.userForm.getRawValue();
 
-    this.authService.register(fullname, email, password, repeatPassword).subscribe(data => {
-      this.router.navigate(['']);
-    });
+    this.authService
+      .register(fullname, email, password, repeatPassword)
+      .subscribe(() => {
+        this.router.navigate(['']);
+      });
   }
 }
